# perf: Merge histograms as sorted slices instead of using temp map