--- a/aggregators/internal/hdrhistogram/sort.go
+++ b/aggregators/internal/hdrhistogram/sort.go
@@ -1,3 +1,7 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License 2.0;
+// you may not use this file except in compliance with the Elastic License 2.0.
+
 package hdrhistogram
 
 import "golang.org/x/exp/constraints"