--- a/aggregators/internal/hdrhistogram/sort_test.go
+++ b/aggregators/internal/hdrhistogram/sort_test.go
@@ -1,0 +1,15 @@
+package hdrhistogram
+
+import (
+	"fmt"
+	"sort"
+)
+
+func Example() {
+	x := []int{3, 2, 1}
+	y := []string{"a", "b", "c"}
+	sort.Sort(SortBy[int, string]{x, y})
+	fmt.Println(x, y)
+	// Output:
+	// [1 2 3] [c b a]
+}